% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cu.hist}
\alias{cu.hist}
\title{Histogram Plotting Function}
\usage{
cu.hist(d, bw = 0, highl = FALSE)
}
\arguments{
\item{d}{the vector of data you want a histogram for}

\item{bw}{optional: set binwidth, otherwise it uses the default from ggplot}

\item{highl}{optional: sets breaks using seq, otherwise uses default}
}
\value{
creates and prints a histogram
}
\description{
Histogram Plotting Function
}
\examples{
cu.hist(d = c(1,1,1,1,2,2,3,3,3,3,3)) # plots a nice histogram of x
cu.hist(d = c(1,1,1,1,2,2,3,3,3,3,3), bw = 200) #plots a histogram of x with binwidth = 200
cu.hist(d = c(1,1,1,1,2,2,3,3,3,3,3), bw = 200, highl = TRUE) #plots a histogram of x with binwidth 200 with the most frequent bin highlighted.
}
