% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{disneyCost}
\alias{disneyCost}
\title{Calculate the cost of a Disney Trip}
\usage{
disneyCost(
  numAdults,
  numDays,
  flightCost,
  inDisney,
  daySpending,
  mealsInPark,
  costOfHomeMeal = 8,
  totalMeals = 3,
  costOfParkMeal = 25,
  inParkHotel = 500,
  outParkHotel = 100
)
}
\arguments{
\item{numDays}{number of days on the trip}

\item{flightCost}{cost of a round trip ticket for one person}

\item{inDisney}{logical indicating whether you will be staying inside the Disney park hotels}

\item{daySpending}{your daily spending limit on souvenirs, etc.}

\item{mealsInPark}{how many meals you will purchase in the park vs. bringing your own food}

\item{costOfHomeMeal}{defaults to 8 dollars, if you anticipate a homemade meal to cost more per person you can adjust this}

\item{totalMeals}{defaults to 3 meals a day, if you plan to eat more meals a day than 3, you can adjust this}

\item{costOfParkMeal}{defaults to 25 dollars, if you anticipate spending more than 25 per person on a meal in the park you can adjust this}

\item{inParkHotel}{defaults to 500 dollars, if you anticipate spending more per night for your in-park hotel, you can adjust this}

\item{outParkHotel}{defaults to 100 dollars, if you anticipate spending more per night for your out-of-park hotel, you can adjust this}

\item{numPeople}{number of people going on the trip over age 2}
}
\value{
a dollar amount for the cost of the trip
}
\description{
Calculate the cost of a Disney Trip
}
\examples{
disneyCost(2, 3, 300, F, 50, 2) # returns the cost of a 3-day trip for 2 adults eating 2 meals in the park, spending $50/day and with a $300 flight per person.

}
