% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{nbaPlot}
\alias{nbaPlot}
\title{Outputs an nba court as a plot}
\usage{
nbaPlot()
}
\value{
graph showing nba court
}
\description{
Outputs an nba court as a plot
}
\examples{
nbaPlot() # shows an NBA court in the Plots window
}
